refractor: add preprocess.c for preprocess-need module
refractor: move import resolution to preprocess.c
